enhancement: Introduce DateTime component
